refactor: improve DX by not requiring signed urls or ssl in development

+++ src/s3Helpers/getFromS3.ts
@@ -2,7 +2,7 @@
 
 import { execSync } from "child_process"
 
-import { SIGNED_GET_URL } from "../constants"
+import { NODE_ENV, SIGNED_GET_URL } from "../constants"
 
 /**
  * A Helper Function to get saved data from s3 by using Signed S3 URLs from env
@@ -11,19 +11,21 @@ import { SIGNED_GET_URL } from "../constants"
  */
 const getFromS3: () => void = () => {
 	// Cleanup old data if any and create the `code` folder
-	execSync(`cd /home/rdamn && rm -rf ./code && runuser -u rdamn -- mkdir /home/rdamn/code`)
+	execSync(`cd /home/rdamn && rm -rf ./code && mkdir ./code && chmod 777 ./code`)
 
-	// Get saved data from s3 and save it as `code.zip`
-	execSync(`runuser -u rdamn -- wget "${SIGNED_GET_URL}" -O /home/rdamn/code/code.zip`)
+	if (NODE_ENV !== "development") {
+		// Get saved data from s3 and save it as `code.zip`
+		execSync(`runuser -u rdamn -- wget "${SIGNED_GET_URL || ""}" -O /home/rdamn/code/code.zip`)
 
-	// Unzip `code.zip` into the `code` folder
-	execSync(`runuser -u rdamn -- unzip -d /home/rdamn/code /home/rdamn/code/code.zip`)
+		// Unzip `code.zip` into the `code` folder
+		execSync(`runuser -u rdamn -- unzip -d /home/rdamn/code /home/rdamn/code/code.zip`)
 
-	// Flatten the top folders of the extracted zip file
-	execSync(`runuser -u rdamn -- mv -f /home/rdamn/code/rdamn-template-*/* /home/rdamn/code/`)
+		// Flatten the top folders of the extracted zip file
+		execSync(`runuser -u rdamn -- mv -f /home/rdamn/code/rdamn-template-*/* /home/rdamn/code/`)
 
-	// Cleanup downloaded files
-	execSync(`runuser -u rdamn -- rm -rf /home/rdamn/code/code.zip /home/rdamn/code/rdamn-template-*`)
+		// Cleanup downloaded files
+		execSync(`runuser -u rdamn -- rm -rf /home/rdamn/code/code.zip /home/rdamn/code/rdamn-template-*`)
+	}
 }
 
 export default getFromS3
